crisp_backend/requirements.txt
django
djangorestframework
psycopg2
firebase-admin
corsheaders